backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0
certifi>=2024.8,<2027.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg[binary]>=3.2,<4.0
PyJWT>=2.9,<3.0
argon2-cffi>=23.1,<24.0
